backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
openai>=1.35.0
google-generativeai>=0.7.2
tavily-python>=0.3.3
redis>=5.0.7
python-dotenv>=1.0.1
pydantic>=2.7.4
httpx>=0.27.0
